+++ src/idc/imgstats/writer/__init__.py
@@ -1,3 +1,4 @@
 from ._area_histogram import AreaHistogramWriter
+from ._contour_areas import ContourAreasWriter
 from ._label_dist import LabelDistributionWriter
 from ._pixel_count import PixelCountWriter