+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - switched to `kasperl` library for base API and generic pipeline plugins
 - added `pixel-count` writer that outputs pixel counts per label for image segmentation annotations
+- added `contour-areas` writer that outputs size and location of contours in image and/or image segmentation annotations
 
 
 0.0.2 (2025-03-14)